Ship DM cloudwatch metrics to hostedgraphite
This is a thin wrapper over crccheck/cloudwatch-to-graphite that allows
us to deploy it on PaaS, configure the metrics we're interested in and
have them sent continuously to hostedgraphite.

To get all our metrics configured in Grafana, these metrics need to
exist, but some metrics may not be present until the first occurence
(e.g. no 5xx responses). In order to mitigate that and still be able to
plot these metrics, we initialize them with a 0 value and a
timestamp of 5 days ago, so as to not conflict with the current data.

Also, because different AWS credentials are needed for preview and
staging/production accounts (to retrieve log metrics) we deploy 2
applications on PaaS, each with its respective credentials.