Drop last metric requested due to incomplete data
When calling leadbutt, we receive metrics from the previous 10 minutes
of data. However, the final minute may not have all the data so
subsequent calls may provide a different value for the same
timestamp. This means that Hosted Graphite will receive two values
for the same metric and the same timestamp. Hosted graphite will
keep both of these values and this may cause inaccuracies when
presenting this data.

We resolve this by stripping off the latest timestamp for each
metric. It will naturally be included in the next leadbutt call and
we will only have one metric value which is accurate.