Send router 429s metric to hosted graphite
If we introduce rate limiting we'll be returning 429s for requests that
have been dropped.

We're not going to send the 429 logs to Kibana (we'll filter them out in
the lambda subscription filter), as one of the reasons we're introducing
IP rate limiting is due to the mad amount of logs we were generating
while being crawled. So we need a way to monitor the number of 429s
we're getting.

A metric has been set up in cloudwatch for the count of 429s, so we can
ship it to hosted graphite and display it on the dashboard, or set up
alerting for it.

This is what this PR is for.