Collect and send 'sum' as well as 'samplecount'
Sum is the better statistic to use, as it allow us to use a default
value for when there's no data. Zero in most cases. Sample count would
include this zero value in its count.

Sending both will allow a transition period. We can stop sending
samplecount after we've switched to sum.