add a "pipe watchdog" which will abort the app if no log messages are received within a certain time threshold
traditional paas "health checks" won't be particularly useful for apps that
listen on no ports when they deadlock. it is probably also preferable to
be performing a watchdog check from outside the main python process to
avoid the possibility of a thread deadlocking while holding the GIL, leaving
the watchdog unable to interject to perform its task.